Removed Embeds.cs, moved all embeds related to campaigns and requests
