+++ CWE/Modules/SuggestionsModule.cs
@@ -1,10 +1,6 @@
 ﻿namespace CWE.Modules
 {
     using System;
-    using System.Collections.Generic;
-    using System.Linq;
-    using System.Text;
-    using System.Threading.Tasks;
     using Discord.Commands;
     using Interactivity;
     using Microsoft.Extensions.Configuration;
