+++ CWE/Services/CommandHandler.cs
@@ -9,6 +9,7 @@
     using CWE.Common;
     using CWE.Data;
     using CWE.Data.Models;
+    using CWE.Modules;
     using Discord;
     using Discord.Addons.Hosting;
     using Discord.Commands;
@@ -92,7 +93,7 @@
                         if (socketMessageComponent.Data.CustomId == "deny")
                         {
                             request.State = RequestState.Denied;
-                            var embed = Embeds.GetRequestEmbed(request);
+                            var embed = RequestsModule.GetRequestEmbed(request);
                             var component = new ComponentBuilder().Build();
 
                             await (socketMessageComponent.Message as IUserMessage).ModifyAsync(x =>
@@ -117,7 +118,7 @@
                         else
                         {
                             request.State = RequestState.Active;
-                            var embed = Embeds.GetRequestEmbed(request);
+                            var embed = RequestsModule.GetRequestEmbed(request);
                             var component = new ComponentBuilder()
                                 .WithButton("Finish", "finish", ButtonStyle.Success)
                                 .Build();
@@ -161,7 +162,7 @@
                         if (socketMessageComponent.Data.CustomId == "finish")
                         {
                             request.State = RequestState.Finished;
-                            var embed = Embeds.GetRequestEmbed(request);
+                            var embed = RequestsModule.GetRequestEmbed(request);
                             var component = new ComponentBuilder().Build();
 
                             await (socketMessageComponent.Message as IUserMessage).ModifyAsync(x =>
@@ -283,7 +284,21 @@
                     var msg = await this.client.GetGuild(this.configuration.GetValue<ulong>("Guild")).GetTextChannel(this.configuration.GetSection("Channels").GetValue<ulong>("Campaigns")).GetMessageAsync(campaign.Message) as IUserMessage;
                     if (msg != null)
                     {
-                        var denied = Embeds.GetDeniedEmbed(campaign);
+                        var denied = new EmbedBuilder()
+                            .WithAuthor(x =>
+                            {
+                                x
+                                .WithIconUrl(Icons.DeniedCampaign)
+                                .WithName("Denied campaign");
+                            })
+                            .AddField("Initiator", $"<@{campaign.Initiator}>", true)
+                            .AddField("Member", $"<@{campaign.User}>", true)
+                            .AddField("Type", campaign.Type, true)
+                            .AddField("Reason", campaign.Reason)
+                            .AddField("Voting", $"This campaign was denied because it didn't receive enough votes within 24 hours.")
+                            .WithColor(Colors.Error)
+                            .Build();
+
                         await msg.ModifyAsync(x => x.Embed = denied);
                         await msg.RemoveAllReactionsAsync();
                     }
@@ -319,7 +334,21 @@
                 return;
             }
 
-            var accepted = Embeds.GetAcceptedEmbed(campaign);
+            var accepted = new EmbedBuilder()
+                .WithAuthor(x =>
+                {
+                    x
+                    .WithIconUrl(Icons.AcceptedCampaign)
+                    .WithName("Accepted campaign");
+                })
+                .AddField("Initiator", $"<@{campaign.Initiator}>", true)
+                .AddField("Member", $"<@{campaign.User}>", true)
+                .AddField("Type", campaign.Type, true)
+                .AddField("Reason", campaign.Reason)
+                .AddField("Voting", $"This campaign received {campaign.Minimal} votes and has thus been accepted.")
+                .WithColor(Colors.Success)
+                .Build();
+
             await msg.ModifyAsync(x => x.Embed = accepted);
             await msg.RemoveAllReactionsAsync();
